line-card-bot/requirements-txt.txt
flask==2.3.3
line-bot-sdk==3.2.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-api-python-client==2.112.0
gspread==5.12.0
filetype==1.2.0
gunicorn==21.2.0
